Update 1.0.1.9, dedimania and mx karma request are now handled by Curl, added karma require finish, admin contact info,fixed mx plugin request

+++ Maps/Maps.php
@@ -872,11 +872,14 @@ class Maps extends ExpPlugin
             } else {
 
                 try {
-                    unlink(Helper::getPaths()->getDefaultMapPath() . $map->fileName);
-                    if ($found) {
-                        $additions = "playlist and disk!";
-                    } else {
-                        $additions = "disk!";
+                    if (file_exists(Helper::getPaths()->getDefaultMapPath() . $map->fileName)) {
+                        unlink(Helper::getPaths()->getDefaultMapPath() . $map->fileName);
+
+                        if ($found) {
+                            $additions = "playlist and disk!";
+                        } else {
+                            $additions = "disk!";
+                        }
                     }
                 } catch (\Exception $ex) {
                     if ($found) {
